Refactored profile to user
