+++ src/server/index.ts
@@ -2,7 +2,7 @@ import express, { Express, Request, Response } from 'express';
 
 const genre = require('./routes/genre');
 
-const profile = require('./routes/profile');
+const profile = require('./routes/user');
 
 const app: Express = express();
 
@@ -13,6 +13,6 @@ app.get('/', (req: Request, res: Response) => {
 });
 
 app.use('/api/genres', genre);
-app.use('/api/profile', profile);
+app.use('/api/user', profile);
 
 module.exports = app;
