+++ src/server/models/userModel.ts
@@ -1,4 +1,4 @@
-export interface Profile {
+export interface User {
   email: string;
   username: string;
   name: string;
